added get_img_link function to views.py
